tests: bsim: bluetooth: Refactor MTU Update test
To isolate Central and Peripheral in two separate builds.
