+++ tests/bsim/bluetooth/host/att/mtu_update/CMakeLists.txt
@@ -5,10 +5,16 @@ cmake_minimum_required(VERSION 3.20.0)
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(bsim_test_mtu_update)
 
-target_sources(app PRIVATE
-  src/main.c
-  ${ZEPHYR_BASE}/samples/bluetooth/mtu_update/central/src/central_mtu_update.c
+target_include_directories(app PRIVATE src)
+
+target_sources_ifdef(CONFIG_BT_PERIPHERAL app PRIVATE
+  src/main_peripheral.c
   ${ZEPHYR_BASE}/samples/bluetooth/mtu_update/peripheral/src/peripheral_mtu_update.c
+)
+
+target_sources_ifdef(CONFIG_BT_CENTRAL app PRIVATE
+  src/main_central.c
+  ${ZEPHYR_BASE}/samples/bluetooth/mtu_update/central/src/central_mtu_update.c
 )
 
 zephyr_include_directories(
