+++ tests/bsim/bluetooth/host/att/mtu_update/compile.sh
@@ -8,6 +8,7 @@ set -eu
 INCR_BUILD=1
 source ${ZEPHYR_BASE}/tests/bsim/compile.source
 
-app="$(guess_test_relpath)" compile
+app="$(guess_test_relpath)" conf_file=prj_central.conf compile
+app="$(guess_test_relpath)" conf_file=prj_peripheral.conf compile
 
 wait_for_background_jobs
