+++ tests/bsim/bluetooth/host/att/mtu_update/prj_central.conf
@@ -1,6 +1,5 @@
 CONFIG_BT=y
 CONFIG_BT_CENTRAL=y
-CONFIG_BT_PERIPHERAL=y
 CONFIG_BT_DEVICE_NAME="MTU Update Test"
 
 CONFIG_BT_GATT_CLIENT=y
